Attempt to prevent dependency cycles
Move handle sensor event from particle.js to sensorAlerts.js
