+++ server/sensorAlerts.js
@@ -95,7 +95,62 @@ async function handleAlert(location, alertType) {
   }
 }
 
+const validateSensorEvent = Validator.body(['coreid', 'event', 'api_key']).exists()
+
+async function handleSensorEvent(request, response) {
+  try {
+    const validationErrors = Validator.validationResult(request).formatWith(helpers.formatExpressValidationErrors)
+
+    if (validationErrors.isEmpty()) {
+      const apiKey = request.body.api_key
+
+      if (particleWebhookAPIKey === apiKey) {
+        let alertType
+        const coreId = request.body.coreid
+        const sensorEvent = request.body.event
+        if (sensorEvent === SENSOR_EVENT.DURATION) {
+          alertType = ALERT_TYPE.SENSOR_DURATION
+        } else if (sensorEvent === SENSOR_EVENT.STILLNESS) {
+          alertType = ALERT_TYPE.SENSOR_STILLNESS
+        } else {
+          const errorMessage = `Bad request to ${request.path}: Invalid event type`
+          helpers.logError(errorMessage)
+        }
+
+        const location = await db.getLocationFromParticleCoreID(coreId)
+        if (!location) {
+          const errorMessage = `Bad request to ${request.path}: no location matches the coreID ${coreId}`
+          helpers.logError(errorMessage)
+          // Must send 200 so as not to be throttled by Particle (ref: https://docs.particle.io/reference/device-cloud/webhooks/#limits)
+          response.status(200).json(errorMessage)
+        } else {
+          if (location.client.isSendingAlerts && location.isSendingAlerts) {
+            await handleAlert(location, alertType)
+          }
+          response.status(200).json('OK')
+        }
+      } else {
+        const errorMessage = `Access not allowed`
+        helpers.logError(errorMessage)
+        // Must send 200 so as not to be throttled by Particle (ref: https://docs.particle.io/reference/device-cloud/webhooks/#limits)
+        response.status(200).json(errorMessage)
+      }
+    } else {
+      const errorMessage = `Bad request to ${request.path}: ${validationErrors.array()}`
+      helpers.logError(errorMessage)
+      // Must send 200 so as not to be throttled by Particle (ref: https://docs.particle.io/reference/device-cloud/webhooks/#limits)
+      response.status(200).json(errorMessage)
+    }
+  } catch (err) {
+    const errorMessage = `Error calling ${request.path}: ${err.toString()}`
+    helpers.logError(errorMessage)
+    // Must send 200 so as not to be throttled by Particle (ref: https://docs.particle.io/reference/device-cloud/webhooks/#limits)
+    response.status(200).json(errorMessage)
+  }
+}
+
 module.exports = {
   setup,
-  handleAlert,
+  handleSensorEvent,
+  validateSensorEvent,
 }
